reconfigure earthfile/docker-compose
Moved 2 Earthfiles (1 for frontend/backend) into a single file and moved
up to the root. Also moved the docker compose up a level. Now you can
use a single command to have the dev environment up and running. Next is
to get the full suite of tests/etc to work from a single file.